+++ tests/testthat/test-vector.R
@@ -1,12 +1,14 @@
-test_that("Test bbox", {
+test_that("Test vector", {
   img <- system.file("img/UK_flag.png", package = "rasterpic")
   x <- sf::st_read(system.file("gpkg/UK.gpkg", package = "rasterpic"),
     quiet = TRUE
   )
 
-  x <- sf::st_bbox(x)
-  expect_s3_class(x, "bbox")
+  x <- as.double(sf::st_bbox(x))
+  expect_true(is.numeric(x))
+  expect_length(x, 4)
 
+  expect_error(rasterpic_img(x[1:3], img))
   expect_message(rasterpic_img(x, img), "'crs' is NA")
   raster <- rasterpic_img(x, img)
 
@@ -24,7 +26,7 @@ test_that("Test bbox", {
   expect_true(terra::xmax(raster) > x[3])
 })
 
-test_that("Test bbox with projs", {
+test_that("Test vector with projs", {
   img <- system.file("img/UK_flag.png", package = "rasterpic")
   x <- sf::st_read(system.file("gpkg/UK.gpkg", package = "rasterpic"),
     quiet = TRUE
@@ -33,8 +35,9 @@ test_that("Test bbox with projs", {
   x_a <- sf::st_transform(x, 25830)
   crs_wkt_sf <- sf::st_crs(x_a)$wkt
 
-  x <- sf::st_bbox(x)
-  expect_s3_class(x, "bbox")
+  x <- as.double(sf::st_bbox(x))
+  expect_true(is.numeric(x))
+  expect_length(x, 4)
 
   raster <- rasterpic_img(x, img, crs = crs_wkt_sf)
   expect_false(terra::crs(raster) == "")