+++ NEWS.md
@@ -6,6 +6,8 @@
 
     -   **sf**: `sfg`, `sf::st_bbox()`.
 
+    -   Numeric vectors `c(xmin, ymin, xmax, ymax)`
+
 # rasterpic 0.1.0
 
 -   Initial release